Баги с удалением и добавление картинок

+++ delete_hotel.php
@@ -15,8 +15,15 @@ $result = mysqli_fetch_row($sql);
 if($result[0] > 0){
     try{
     mysqli_query($con,"DELETE FROM `likes` WHERE id_hotels = $deleteHot "); 
+    $hotel_image = fetchAll($con,"SELECT `image`, `image2`, `image3`,`image4` FROM hotel_image WHERE id_hotel = $deleteHot");
+   // print_r($hotel_image);
+    mysqli_query($con,"DELETE FROM `hotel_image` WHERE `id_hotel` = $deleteHot "); 
+    mysqli_query($con,"DELETE FROM `comments` WHERE `hotel_id` = $deleteHot "); 
 	mysqli_query($con,"DELETE FROM `hotels` WHERE `user_id` = $username[id] AND id = $deleteHot "); 
-    unlink($deleteImg);
+    if(file_exists($deleteImg)){
+        unlink($deleteImg);
+    }
+
     header("Location: /");
     exit();
     }
